+++ src/factory/DirectiveFactory.js
@@ -22,14 +22,18 @@ Util.ext(_DirectiveFactory.prototype,{
 	hasEndTag : function(type){
 		return this.types[type].props.$endTag;
 	},
-
 	newInstanceOf : function(type,node,component,attrName,attrValue){
 		if(!this.types[type])return null;
 
 		var rs = new this.types[type].clazz(this.baseClass,attrName,attrValue,component);
 		Util.extProp(rs,this.types[type].props);
 
-		rs.$view = new View(node);
+		if(node.__impex__view){
+			rs.$view = node.__impex__view;
+		}else{
+			rs.$view = new View(node);
+			node.__impex__view = rs.$view;
+		}
 		
 		//inject
 		var services = null;